+++ app/src/main/java/com/enaboapps/switchify/utils/StringUtils.kt
@@ -5,7 +5,7 @@ class StringUtils {
     companion object {
 
         // Function to get time string in seconds from milliseconds (ex. 1.5 seconds)
-        fun getSecondsString(milliseconds: Int): String {
+        fun getSecondsString(milliseconds: Long): String {
             val seconds: Double = (milliseconds / 1000.0)
             return "$seconds seconds"
         }
